requirements.txt
# Full stack: API, dashboard, DB scripts, training, and tests (Python 3.11+).
python-dotenv>=1.0.0
psycopg[binary]>=3.2.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.36
pandas>=2.2.0
numpy>=1.26.0
Faker>=24.0

fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
anthropic>=0.40.0

joblib>=1.4.0
xgboost>=2.1.0
scikit-learn>=1.5.0
mlflow>=2.13.0

pytest>=8.3.0

streamlit>=1.40.0
folium>=0.18.0
streamlit-folium>=0.24.0
